disable problematic test
